refactor tests and updated route
